move last components to modules
