+++ src/modules/navbar/settings-dropdown.vue
