+++ src/modules/player/progress-bar.vue
