+++ src/modules/player/volume-slider.vue
